+++ lib/virtus/matchers/integrations.rb
@@ -1,5 +1,0 @@
-if defined?(RSpec)
-  RSpec.configure do |config|
-    config.include Virtus::Matchers, type: :virtus
-  end
-end