+++ lib/virtus/matchers/rspec.rb
@@ -1,0 +1,5 @@
+require "virtus/matchers"
+
+RSpec.configure do |config|
+  config.include Virtus::Matchers, type: :virtus
+end